+++ frontend/src/components/common/TripItem/TripItem.tsx
@@ -1,38 +1,21 @@
-import MoreIcon from '@assets/svg/more-icon.svg';
 import { CURRENCY_ICON } from '@constants/trip';
 import type { TripItemData } from '@type/tripItem';
-import {
-  Box,
-  Flex,
-  Heading,
-  ImageCarousel,
-  Menu,
-  MenuItem,
-  MenuList,
-  Text,
-  Theme,
-  Toast,
-  useOverlay,
-} from 'hang-log-design-system';
+import { Box, Flex, Heading, ImageCarousel, Text, Theme } from 'hang-log-design-system';
 
 import { formatNumberToMoney } from '@utils/formatter';
 
-import { useDeleteTripItemMutation } from '@hooks/api/useDeleteTripItemMutation';
 import { useDraggedItem } from '@hooks/common/useDraggedItem';
 
 import StarRating from '@components/common/StarRating/StarRating';
+import EditMenu from '@components/common/TripItem/EditMenu/EditMenu';
 import {
   expenseStyling,
   getContainerStyling,
   informationContainerStyling,
   memoStyling,
-  moreButtonStyling,
-  moreMenuListStyling,
-  moreMenuStyling,
   starRatingStyling,
   subInformationStyling,
 } from '@components/common/TripItem/TripItem.style';
-import TripItemAddModal from '@components/trip/TripItemAddModal/TripItemAddModal';
 
 interface TripListItemProps extends TripItemData {
   tripId: number;
@@ -50,23 +33,7 @@ const TripItem = ({
   onDragEnd,
   ...information
 }: TripListItemProps) => {
-  const deleteTripItemMutation = useDeleteTripItemMutation();
-  const { isOpen: isMenuOpen, open: openMenu, close: closeMenu } = useOverlay();
-  const { isOpen: isEditModalOpen, open: openEditModal, close: closeEditModal } = useOverlay();
-  const { isOpen: isErrorTostOpen, open: openErrorToast, close: closeErrorToast } = useOverlay();
   const { isDragging, handleDrag, handleDragEnd } = useDraggedItem(onDragEnd);
-
-  const handleTripItemDelete = () => {
-    deleteTripItemMutation.mutate(
-      { tripId, itemId: information.id },
-      {
-        onError: () => {
-          openErrorToast();
-          closeMenu();
-        },
-      }
-    );
-  };
 
   return (
     // ! 수정 모드에서만 drag할 수 있다
@@ -113,48 +80,8 @@ const TripItem = ({
           </Box>
         </Flex>
         {/* ! 로그인 + 수정 모드일 떄만 볼 수 있다 */}
-        <Menu css={moreMenuStyling} closeMenu={closeMenu}>
-          <button css={moreButtonStyling} type="button" onClick={openMenu}>
-            <MoreIcon />
-          </button>
-          {isMenuOpen && (
-            <MenuList css={moreMenuListStyling}>
-              <MenuItem onClick={openEditModal}>수정</MenuItem>
-              <MenuItem onClick={handleTripItemDelete}>삭제</MenuItem>
-            </MenuList>
-          )}
-        </Menu>
-        {isEditModalOpen && (
-          <TripItemAddModal
-            tripId={tripId}
-            itemId={information.id}
-            dayLogId={dayLogId}
-            onClose={closeEditModal}
-            initialData={{
-              itemType: information.itemType,
-              dayLogId,
-              title: information.title,
-              isPlaceUpdated: false,
-              place: null,
-              rating: information.rating,
-              expense: information.expense
-                ? {
-                    currency: information.expense.currency,
-                    amount: information.expense.amount,
-                    categoryId: information.expense.category.id,
-                  }
-                : null,
-              memo: information.memo,
-              imageUrls: information.imageUrls,
-            }}
-          />
-        )}
+        <EditMenu tripId={tripId} dayLogId={dayLogId} {...information} />
       </li>
-      {isErrorTostOpen && (
-        <Toast variant="error" closeToast={closeErrorToast}>
-          아이템 삭제에 실패했습니다.
-        </Toast>
-      )}
     </>
   );
 };
